# perf: Parallelize child modules for env_var|custom