Top modules are processed in parallel, but child modules (of env_var|custom) are not. This PR fixed that.

#### Description

Given a config like this:
```
format = """$custom"""
command_timeout = 3000

[custom.foo]
command = 'sleep 1; echo foo'
when = 'true'

[custom.bar]
command = 'sleep 1; echo bar'
when = 'true'
```

The two custom modules are processed in sequence and takes ~2s:

```
$ time STARSHIP_CONFIG=~/.config/test.toml starship prompt

%{%}foo bar %{%}STARSHIP_CONFIG=~/.config/test.toml starship prompt  0.02s user 0.02s system 1% cpu 2.058 total
```

After parallelization, it takes ~1s, as expected:

```
$ time STARSHIP_CONFIG=~/.config/test.toml ./target/release/starship prompt

%{%}foo bar %{%}STARSHIP_CONFIG=~/.config/test.toml ./target/release/starship prompt  0.02s user 0.03s system 4% cpu 1.038 total
```

Note that if we explicitly declare them as top modules like this:

```
format = """${custom.foo}${custom.bar}"""
```

The existing parrelization works as expected.

#### How Has This Been Tested?
<!--- Please describe in detail how you tested your changes. -->
<!--- Include details of your testing environment, tests ran to see how -->
<!--- your change affects other areas of the code, etc. -->
- [x] I have tested using **MacOS**
- [ ] I have tested using **Linux**
- [ ] I have tested using **Windows**

#### Checklist:
<!--- Go over all the following points, and put an `x` in all the boxes that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're here to help! -->
- [ ] I have updated the documentation accordingly.
- [ ] I have updated the tests accordingly.